A functional API with basic user operations
Tested, works fine.
